requirements.txt
scipy==1.5.4
torchvision==0.8.2+cu101
opencv_python==4.5.1.48
numpy==1.19.5
torch==1.7.1+cu101
Pillow==8.1.2
tensorboardX==2.1
